Have matching capture all binding values.
This will make consistent variables the default. Also, variables in
slurps are matched with the sequence of arguments rather than each
individually (as described in #5)

This should allow destructuring to use the captured values rather than
having to walk the pattern tree again, <s>hopefully making the process
faster</s> this is currently slower, but can be made faster.